ai-ml/requirements.txt
Flask==3.1.2
flask-cors==6.0.1
pdfminer.six==20231228
requests==2.32.3

# Core scientific stack
numpy==2.1.3
scipy==1.14.1
pandas==2.2.3
joblib==1.4.2

# Classical ML
scikit-learn==1.5.2

# Optional / heavier components - install when needed
# sentence-transformers and torch are optional for embeddings
sentence-transformers==2.2.2
torch==2.2.0
transformers==4.45.2

# spaCy (NER)
spacy==3.7.5
